% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kmeans}
\alias{kmeans}
\title{kmeans using yakmo}
\usage{
kmeans(X, k, verbose = FALSE)
}
\arguments{
\item{X}{data matrix}

\item{k}{number of clusters}

\item{verbose}{verbose output?}
}
\description{
kmeans using yakmo
}

