% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yakmoKMeans.R
\name{orthoKMeansPredict}
\alias{orthoKMeansPredict}
\title{orthogonal kmeans prediction function}
\usage{
orthoKMeansPredict(x, rounds = 1, verbose = FALSE)
}
\arguments{
\item{x}{data to assign clusters}

\item{rounds}{list of views, if multiple views were computed}

\item{verbose}{show verbose messages?}
}
\value{
numericmatrix with as many colums as specified in the rounds array
}
\description{
orthogonal kmeans prediction function
}

