% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{orthoKMeansPredictCpp}
\alias{orthoKMeansPredictCpp}
\title{K-Means prediction using yakmo library}
\usage{
orthoKMeansPredictCpp(x, centers, verbose = FALSE)
}
\arguments{
\item{x}{data matrix}

\item{centers}{centers FIXME}

\item{verbose}{verbose output?}
}
\value{
a list consisting of
	centers	these are the resulting centroids of the kmean algorithm
	cluster 	these are the labels for the resulting clustering
	obj			this is a vector with the final objective value for each round
	dim			dimension of the input space (=dim of centroids)
	allcenters	this is the list of centroids, one matrix of centroids for each round
	allcluster		this is the list of labels, one vector for each round
}
\description{
K-Means prediction using yakmo library
}

