% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yakmoKMeans.R
\name{orthoKMeansTrain}
\alias{orthoKMeansTrain}
\title{orthogonal kmeans training function}
\usage{
orthoKMeansTrain(x = NULL, k = 8, rounds = 1, iter.max = 100L,
  init.type = "Random", verbose = FALSE)
}
\arguments{
\item{x}{data to cluster}

\item{k}{number of centroids}

\item{rounds}{number of rounds/views for orthogonal kmeans}

\item{iter.max}{number of maximal iterations for each clustering}

\item{init.type}{string with method to initialize centroids}

\item{verbose}{show verbose messages?}
}
\description{
orthogonal kmeans training function
}

