% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{orthoKMeansTrainCpp}
\alias{orthoKMeansTrainCpp}
\title{K-Means using yakmo library}
\usage{
orthoKMeansTrainCpp(x = matrix(), k = 3L, iter = 100L, rounds = 1L,
  initType = "Random", random = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{data matrix}

\item{k}{number of clusters}

\item{iter}{numer of iterations in one round}

\item{rounds}{number of rounds (orthogonal views)}

\item{initType}{centroid initialization via Random or KMeans++}

\item{random}{use random or pseudo-random (seeded) generator?}

\item{verbose}{verbose output?}
}
\value{
a list consisting of
	centers	these are the resulting centroids of the kmean algorithm (as a std::vector of NumericMatrix)
	cluster 	these are the labels for the resulting clustering (as a std::vector of NumericVector)
	obj			this is a vector with the final objective value for each round
}
\description{
K-Means using yakmo library
}

