% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yakmoR.R
\docType{package}
\name{yakmoR}
\alias{yakmoR}
\alias{yakmoR-package}
\title{yakmoR: A package for fast K-Means with multi-view clustering via orthogonalization.}
\description{
The yakmoR package is a wrapper for the yakmo library. It uses basic K-Means, and provides
multi-view clustering via orthogonalization as well as an option to use KMeans++ initialization.
}
\section{yakmoR functions}{

orthoKMeansTrain
orthoKMeansPredict
}

