% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/orthoKMeansPredict.R
\name{orthoKMeansPredict}
\alias{orthoKMeansPredict}
\title{orthogonal kmeans prediction function}
\usage{
orthoKMeansPredict(x, obj = NULL, verbose = FALSE)
}
\arguments{
\item{x}{data to assign clusters}

\item{obj}{an object returned by orthoKMeansTrain}

\item{verbose}{show verbose messages?}
}
\value{
a matrix with as many colums as rounds trained
}
\description{
orthogonal kmeans prediction function
}
\examples{
obj = yakmoR::orthoKMeansTrain (x = as.matrix(iris[seq(1,150,2),1:4]),
			k = 3, rounds = 3)
	predictions = yakmoR::orthoKMeansPredict (x = as.matrix(iris[seq(2, 150, 2),1:4]),
			obj = obj)
}

